requirements.txt
supabase==1.0.3
pandas==2.0.3
python-dotenv==1.0.0
